requirements.txt
screeninfo~=0.8.1
pillow~=10.2.0
requests~=2.31.0
schedule~=1.2.1
bs4~=0.0.2
beautifulsoup4~=4.12.3
pywin32
lxml